Update README.md
(1). I change the style as "Question - Answer". Then, we can clearly find each question we need to answer and fill free to add/erase answers to the questions. But we will change it as a formal proposal report before we submit it. 
(2). About Qeustion-(b): The static analysis qeustion that your tool will answer? 
Based on the example Joseph wrote down, it should belongs to "operator covnersions", but I don't want to only mention this issue. It will narrow down our goal (In case we found it is very difficult to solve in the following days). So I add several more questions such as "casting operations" etc. We don't need to cover so many topics but only choose one or two.